requirements.txt
pytest==7.4.0
pandas==2.2.0
pyarrow==15.0.0
python-dotenv==1.0.0
requests==2.31.0
ujson==1.0.0
